Remove indentation from JSDoc descriptions in tests. (...)
Somewhere between 3.8 and 5.0, indentation for JSDoc description lines starting
with whitespace was fixed to not include the additional space immediately after
the line's leading `*`. The indentiation in this test causes it to fail in at
least TS 5.0, so I'm removing it rather than trying to track down exactly when
this change was made and then propagating the version number to the test.